requirements.txt
mujoco
numpy==1.24.4
numpy_quaternion==2023.0.3
open3d==0.13.0
opencv_python==4.9.0.80
spatialmath_python==1.1.9
glfw==2.7.0
pytest==8.3.3

tensorboard==2.14.0
torch==2.3.0
dm_control==1.0.23
pdoc==14.7.0